refactor: Split lottery hooks
