+++ src/views/Lottery/components/BuyTicketsButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Button, useModal, WaitIcon, ButtonProps } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { LotteryStatus } from 'config/constants/types'
 import BuyTicketsModal from './BuyTicketsModal/BuyTicketsModal'
