+++ src/views/Lottery/components/BuyTicketsModal/BuyTicketsModal.tsx
@@ -20,7 +20,8 @@ import { getFullDisplayBalance } from 'utils/formatBalance'
 import { getCakeAddress } from 'utils/addressHelpers'
 import { BIG_ZERO, ethersToBigNumber } from 'utils/bigNumber'
 import { useAppDispatch } from 'state'
-import { usePriceCakeBusd, useLottery } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { fetchUserTicketsAndLotteries } from 'state/lottery'
 import useTheme from 'hooks/useTheme'
 import useTokenBalance, { FetchStatus } from 'hooks/useTokenBalance'
