+++ src/views/Lottery/components/ClaimPrizesModal/ClaimPrizesInner.tsx
@@ -5,7 +5,8 @@ import { useTranslation } from 'contexts/Localization'
 import { LotteryTicket, LotteryTicketClaimData } from 'config/constants/types'
 import { getBalanceAmount } from 'utils/formatBalance'
 import { callWithEstimateGas } from 'utils/calls'
-import { useLottery, usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { fetchUserLotteries } from 'state/lottery'
 import { useAppDispatch } from 'state'
 import Balance from 'components/Balance'
