+++ src/views/Lottery/components/ClaimPrizesModal/index.tsx
@@ -7,7 +7,7 @@ import { delay } from 'lodash'
 import confetti from 'canvas-confetti'
 import { LotteryTicketClaimData } from 'config/constants/types'
 import { useAppDispatch } from 'state'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { fetchUserLotteries } from 'state/lottery'
 import ClaimPrizesInner from './ClaimPrizesInner'
