+++ src/views/Lottery/components/Hero.tsx
@@ -3,7 +3,8 @@ import styled, { keyframes } from 'styled-components'
 import { Box, Flex, Heading, Skeleton } from '@pancakeswap/uikit'
 import { LotteryStatus } from 'config/constants/types'
 import { useTranslation } from 'contexts/Localization'
-import { useLottery, usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { getBalanceNumber } from 'utils/formatBalance'
 import Balance from 'components/Balance'
 import { TicketPurchaseCard } from '../svgs'
