+++ src/views/Lottery/components/PreviousRoundCard/Footer.tsx
@@ -3,7 +3,8 @@ import styled from 'styled-components'
 import { Flex, ExpandableLabel, CardFooter, Skeleton, Heading, Box, Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { LotteryRound } from 'state/types'
-import { useGetLotteryGraphDataById, usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd } from 'state/hooks'
+import { useGetLotteryGraphDataById } from 'state/lottery/hooks'
 import { formatNumber, getBalanceNumber } from 'utils/formatBalance'
 import Balance from 'components/Balance'
 import RewardBrackets from '../RewardBrackets'
