+++ src/views/Lottery/components/ViewTicketsModal/CurrentRoundTicketsInner.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Flex, Box, Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { LotteryStatus } from 'config/constants/types'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import useTheme from 'hooks/useTheme'
 import TicketNumber from '../TicketNumber'
 import BuyTicketsButton from '../BuyTicketsButton'
