+++ src/views/Lottery/components/ViewTicketsModal/PreviousRoundTicketsInner.tsx
@@ -18,7 +18,7 @@ import { LotteryTicket, LotteryTicketClaimData } from 'config/constants/types'
 import { fetchLottery, fetchTickets } from 'state/lottery/helpers'
 import { getWinningTickets } from 'state/lottery/fetchUnclaimedUserRewards'
 import { LotteryRound } from 'state/types'
-import { useGetUserLotteryGraphRoundById } from 'state/hooks'
+import { useGetUserLotteryGraphRoundById } from 'state/lottery/hooks'
 import { useTranslation } from 'contexts/Localization'
 import useTheme from 'hooks/useTheme'
 import WinningNumbers from '../WinningNumbers'
