+++ src/views/Lottery/components/ViewTicketsModal/index.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Modal } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { LotteryStatus } from 'config/constants/types'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import useTheme from 'hooks/useTheme'
 import PreviousRoundTicketsInner from './PreviousRoundTicketsInner'
 import CurrentRoundTicketsInner from './CurrentRoundTicketsInner'
