+++ src/views/Lottery/components/YourHistoryCard/FinishedRoundTable.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Text, Box, Flex } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { LotteryStatus } from 'config/constants/types'
-import { useGetUserLotteriesGraphData } from 'state/hooks'
+import { useGetUserLotteriesGraphData } from 'state/lottery/hooks'
 import FinishedRoundRow from './FinishedRoundRow'
 
 const Grid = styled(Box)`
