+++ src/views/Lottery/hooks/useNextEventCountdown.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState, useRef } from 'react'
 import { useAppDispatch } from 'state'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { fetchCurrentLottery, setLotteryIsTransitioning } from 'state/lottery'
 
 const useNextEventCountdown = (nextEventTime: number): number => {
