+++ src/views/Lottery/hooks/useStatusTransitions.tsx
@@ -3,7 +3,7 @@ import { LotteryStatus } from 'config/constants/types'
 import usePreviousValue from 'hooks/usePreviousValue'
 import { useEffect } from 'react'
 import { useAppDispatch } from 'state'
-import { useLottery } from 'state/hooks'
+import { useLottery } from 'state/lottery/hooks'
 import { fetchPublicLotteries, fetchCurrentLotteryId, fetchUserLotteries } from 'state/lottery'
 
 const useStatusTransitions = () => {
