+++ src/views/Lottery/index.tsx
@@ -5,7 +5,7 @@ import { LotteryStatus } from 'config/constants/types'
 import PageSection from 'components/PageSection'
 import { useTranslation } from 'contexts/Localization'
 import useTheme from 'hooks/useTheme'
-import { useFetchLottery, useLottery } from 'state/hooks'
+import { useFetchLottery, useLottery } from 'state/lottery/hooks'
 import {
   TITLE_BG,
   GET_TICKETS_BG,
